+++ src/preload.js
@@ -1,0 +1,4 @@
+const remote = require('@electron/remote');
+const mainWindow = remote.getCurrentWindow();
+
+console.log(mainWindow);