+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## WIP
+
+### Added
+- Window title
+
+### Fixes
+- Remove hour from timer
+
+---
+
 ## [0.1.0] - 2021-05-25
 
 ### Added